feat(ban)!: Merge ban until and ban with message commands into one.
Merges the `/ban until` and `/ban with-message` commands into one `/ban`
command. Additionally separates the ban reason and message into separate
parameters, so that the internal info can contain details one would not
want to send to the user being banned. Also adds a parameter to delete
recent messages.
